requirements.txt
fastapi
uvicorn
ccxt
python-dotenv
